VMMakerJS-bf.16: Pointer arithmetic for Balloon2D and Genie
Name: VMMakerJS-bf.16
Author: bf
Time: 13 November 2014, 7:08:06.162 pm
UUID: aa9fd3da-7901-4c62-ae34-c166f1492d3b
Ancestors: VMMakerJS-bf.15

Balloon2D needs pointer arithmetic. Fake it.
How? If p is a 'pointer' to an Int32Array, then
	p += offset
becomes
	p = new Int32Array(p.buffer, p.byteOffset + 4 * offset)
This is actually more efficient than it looks. The Int32Array is just
an object holding a buffer, and offset. Allocating this is cheap, and
it's going to be GC'ed quickly. The resulting plugin is actually twice
as fast as my home-grown one.