Repair touch event interpretation on Android Firefox.
Android Firefox (98, probably earlier too) is able to use
`offsetX`/`offsetY` in mouse events, despite not having "AppleWebKit" in
its user agent string. This patch first tries checking for
"AppleWebKit", as before, but if it is not found, delays a decision
until later, when the event is checked for an actual `offsetX` field.